@@ -9,11 +9,12 @@
     "@testing-library/user-event": "^13.5.0",
     "ajv": "^8.12.0",
     "ajv-keywords": "^5.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "8.1.3",
-    "redux": "4.2.1",
     "react-scripts": "5.0.1",
+    "redux": "4.2.1",
+    "sass": "^1.83.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
